+++ numerical-integration/src/integration/left-riemann-sum.c
@@ -1,11 +1,11 @@
 #include "integration/integration-methods.h"
 
-double left_riemann_sum(function func, double a, double b, int n) {
-    double h = (b - a) / n;
-    double sum = 0;
+long double left_riemann_sum(function func, long double a, long double b, int n) {
+    long double h = (b - a) / n;
+    long double sum = 0;
 
     for (int i = 0; i < n; i++) {
-        double x = a + i * h;
+        long double x = a + i * h;
         sum += func(x);
     }
